sort shell customisation
