+++ ganga/GangaCore/Runtime/bootstrap.py
@@ -1130,8 +1130,7 @@ under certain conditions; type license() for details.
             enableCaching()
 
             self.check_IPythonDir()
-            self.launch_IPython(local_ns, GangaProgram._ganga_error_handler, self.ganga_prompt)
-
+            self.launch_IPython(local_ns)
         else:
             override_credits()
             import code
@@ -1187,32 +1186,9 @@ under certain conditions; type license() for details.
 
         return None
 
-    @staticmethod
-    def _ganga_error_handler(exception_obj, etype, value, tb, tb_offset=None):
-        """
-        Error handler for IPython 3.x+ to identify expected Ganga exceptions or unexpected uncaught exceptions from somewhere
-        """
-        ## see https://ipython.org/ipython-doc/dev/api/generated/IPython.core.interactiveshell.html#IPython.core.interactiveshell.InteractiveShell.set_custom_exc
-        from GangaCore.Utility.logging import getLogger
-        logger = getLogger()
-        #logger.error("%s" % value)
-
-        from GangaCore.Core.exceptions import GangaException
-        import traceback
-        # Extract the stack from this traceback object
-        stack = traceback.extract_tb(tb)
-        # If this is an error from the interactive prompt then the length is 2, otherwise the error is from deeper in Ganga
-        if not issubclass(etype, GangaException) and len(stack) > 2:
-            logger.error("!!Unknown/Unexpected ERROR!!")
-            logger.error("If you're able to reproduce this please report this to the Ganga developers!")
-            #logger.error("value: %s" % value)
-            exception_obj.showtraceback((etype, value, tb), tb_offset=tb_offset)
-        else:
-            logger.error("%s" % value)
-        return None
 
     @staticmethod
-    def launch_IPython(local_ns, error_handler, ganga_prompt):
+    def launch_IPython(local_ns):
         """
         Launch an embedded IPython session within the GangaCore.GPI namespace
         """
@@ -1233,50 +1209,16 @@ under certain conditions; type license() for details.
         banner = exit_msg = ''
         cfg.TerminalInteractiveShell.prompts_class = GangaPrompt
         cfg.TerminalInteractiveShell.banner1 = banner
+
         # Setting up the confirmation of exit on Ctrl+D this prompt annoys some users(developers)
         from GangaCore.Utility.Config.Config import getConfig
         config = getConfig('Configuration')
         cfg.InteractiveShell.confirm_exit = config['confirm_exit']
+        #The customisation of the interactive shell is done in this extension
+        cfg.InteractiveShellApp.extensions = ['GangaCore.Runtime.ganga_extension']
 
-        from IPython import start_ipython, get_ipython
-        """
-        ipshell.events.register("post_execute", ganga_prompt)
-
-        # Add our custom error handler to ignore stack traces for GangaExceptions
-        ipshell.set_custom_exc((Exception,), error_handler)
-
-        # buffering of log messages from all threads called "GANGA_Update_Thread"
-        # the logs are displayed at the next IPython prompt
-        from GangaCore.Utility.logging import enableCaching
-        enableCaching()
-
-        # Magic ganga function for replacing 'execfile'
-        from GangaCore.Runtime.IPythonMagic import GangaMagics
-        ipshell.register_magics(GangaMagics)
-
-        """
-
-        start_ipython(argv = [], user_ns = local_ns, local_ns=local_ns, module=GangaCore.GPI, config = cfg)
-    @staticmethod
-    def ganga_prompt(_=None):
-        #Flush the logging
-        from GangaCore.Utility.logging import flushAtIPythonPrompt
-        flushAtIPythonPrompt()
-
-        try:
-            from GangaCore.GPIDev.Credentials import get_needed_credentials
-
-            needed_credentials = get_needed_credentials()
-
-            # Add still-needed credentials to the prompt
-            if needed_credentials:
-                prompt = 'Warning, some credentials needed by the monitoring are missing or invalid:\n'
-                for cred_req in needed_credentials:
-                    prompt += '  ' + str(cred_req).replace('\n ', '') + '\n'
-                prompt += 'Call `credential_store.renew()` to update them.\n'
-                print(prompt)
-        except KeyboardInterrupt:
-            return
+        import IPython
+        IPython.start_ipython(argv = [], user_ns = local_ns, local_ns=local_ns, module=GangaCore.GPI, config = cfg)
 
 
 #
